requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
decorator==4.4.2
Django==2.2.4
django-appconf==1.0.5
django-crispy-forms==1.9.2
django-etc==1.3.1
django-hitcount==1.3.3
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
moviepy==1.0.3
mysqlclient==1.4.6
numpy==1.22.1
pilkit==2.0
Pillow==9.0.0
proglog==0.1.9
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.8
